// src/types/enum_routes.ts
export enum FrontEndRoutes {
	ADDMEAL = '/addmeal',
	ADMIN = '/admin',
	BASE = '/',
	CATCH_ALL = '/:catchAll(.*)',
	EDITMEAL = '/editmeal',
	ERROR = '/error',
	FLUSH_CACHE='/flushcache',
	FORGOTPASSWORD = '/forgotpassword',
	MEALS = '/meals',
	REGISTER = '/register',
	RESETPASSWORD_param_ID = '/user/reset/:id',
	SETTINGS = '/settings',
	SIGNIN = '/signin',
	USER_RESET = '/user/reset',
	USER_VERIFY_param_ID = '/user/verify/:id',
}

export enum FrontEndNames {
	ADDMEAL = 'addmeal',
	ADMIN = 'admin',
	EDITMEAL = 'editmeal',
	ERROR = 'error',
	FLUSH_CACHE = 'flush_cache',
	FORGOTPASSWORD = 'forgot-password',
	HOME = 'home',
	REGISTER = 'register',
	SETTINGS = 'settings',
	SIGNIN = 'signin',
	USER_RESET = 'user_reset',
	USER_RESET_ID = 'user_reset_id',
	USER_VERIFY_param_ID = 'user_verify_id',
}